% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools4CommonMetabolomics.R
\name{calculateExactMass}
\alias{calculateExactMass}
\title{calculateExactMass}
\usage{
calculateExactMass(formula)
}
\arguments{
\item{formula}{chemical formula}
}
\description{
calculate exact mass
}
\examples{
calculateExactMass(formula = "C2H5OH")
}
\author{
Zhiwei Zhou
}
