% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools4InsilicoMsMs.R
\name{convertAdduct4Sirius}
\alias{convertAdduct4Sirius}
\title{convertAdduct4Sirius}
\usage{
convertAdduct4Sirius(adduct, source = "MetDNA2")
}
\arguments{
\item{adduct}{adduct type need to convert. e.g., '\link{M+H}+', '\link{M-H+2Na}+'}

\item{source}{'MetDNA2'}
}
\description{
convert adduct type to fit MS-FINDER (Default: MSFINDER V3.24)
}
\examples{
convertAdduct4MsFinder(adduct = '[M-H+2Na]+')
}
\author{
Zhiwei Zhou
}
