% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools4InsilicoMsMs.R
\name{convertAdduct4Tools}
\alias{convertAdduct4Tools}
\title{convertAdduct4Tools}
\usage{
convertAdduct4Tools(
  adduct,
  tool_from = c("metdna2", "msfinder", "sirius"),
  tool_to = c("metdna2", "msfinder", "sirius")
)
}
\arguments{
\item{adduct}{adduct type}

\item{tool_from}{annotation tool name, including: 'metdna2', 'msfinder', 'sirius'. Default: 'metdna2'}

\item{tool_to}{annotation tool name, including:  'metdna2', 'msfinder', 'sirius'. Default: 'metdna2'}
}
\description{
convert adduct types between different in-silico annotation tools
}
\examples{
# single adduct
convertAdduct4Tools(adduct = '[M+H]+', tool_from = 'metdna2', tool_to = 'sirius')
# multiple adducts
convertAdduct4Tools(adduct = c('[M+H]+', '[M-H+2Na]+'), tool_from = 'metdna2', tool_to = 'msfinder')
}
\author{
Zhiwei Zhou
}
