% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools4CommonMetabolomics.R
\name{convertMz2Adduct}
\alias{convertMz2Adduct}
\title{convertMz2Adduct}
\usage{
convertMz2Adduct(base_mz, base_adduct, ...)
}
\arguments{
\item{base_mz}{numeric}

\item{base_adduct}{character}

\item{...}{Other arguments passed on to \code{\link[=transformMz]{transformMz()}}}
}
\description{
mz conversion between different adduct forms
}
\examples{
convertMz2Adduct(base_mz = 181.0707,
                 base_adduct = '[M+H]+',
                 adduct_list = NULL,
                 type = 'adduct',
                 polarity = 'positive')

convertMz2Adduct(base_mz = 181.0707,
                 base_adduct = '[M+H]+',
                 adduct_list = c('[M-H2O+H]+', '[M+Na]+'),
                 type = 'nl',
                 polarity = 'negative')

convertMz2Adduct(base_mz = 143.0347,
                 base_adduct = '[3M-H]-',
                 adduct_list = c('[M-H2O+H]+', '[M+Na]+'))
}
\author{
Zhiwei ZHou
}
