% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools4CommonMetabolomics.R
\name{convertStructureFormat}
\alias{convertStructureFormat}
\title{convertStructureFormat}
\usage{
convertStructureFormat(
  file_input,
  file_output,
  format_input,
  format_output,
  options = ""
)
}
\arguments{
\item{format_input}{See \link{https://open-babel.readthedocs.io/en/latest/FileFormats/Overview.html#file-formats}}

\item{format_output}{See \link{https://open-babel.readthedocs.io/en/latest/FileFormats/Overview.html#file-formats}}
}
\description{
R wrapper for OpebBabel
}
\author{
Zhiwei Zhou
}
