% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools4InsilicoMsMs.R
\name{copyMetDNA2}
\alias{copyMetDNA2}
\title{copyMetDNA2}
\usage{
copyMetDNA2(path_from = ".", path_to = ".", dir_ignore = NULL)
}
\arguments{
\item{path_from}{'.'}

\item{path_to}{'.'}

\item{dir_ignore}{file names for ignore}
}
\description{
copyMetDNA2
}
\examples{
copyMetDNA2(path_from = 'Z:/memberdata/ZHOUzhiwei/01_data/20210517_nist_urine_extended_steps/dp_pos_step0',
            path_to = 'H:/00_projects/03_MetDNA2/00_data/20210519_mutiple_ext_step_nist_urine',
            dir_ignore = '02_formula_prediction')
}
\author{
Zhiwei Zhou
}
