% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools4PublicDB.R
\name{ExtractPropertyFromHMDB}
\alias{ExtractPropertyFromHMDB}
\title{extractPropertyFromHMDB}
\usage{
ExtractPropertyFromHMDB(id_hmdb)
}
\arguments{
\item{id_hmdb}{Identifier of human metabolome database}
}
\value{
a data.frame of properties
}
\description{
Extract properties from HMDB
}
\examples{
test <- extractPropertyFromHMDB(id_hmdb = 'HMDB0000001')
}
\author{
Zhiwei Zhou
}
