% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools4PublicDB.R
\name{extractPropertyFromPubChem}
\alias{extractPropertyFromPubChem}
\title{extractPropertyFromPubChem}
\usage{
extractPropertyFromPubChem(id_pubchem)
}
\arguments{
\item{id_pubchem}{Identifier of pubchem}
}
\value{
a data.frame of properties
}
\description{
Extract properties from Pubchem
}
\examples{
test <- extractPropertyFromPubChem(id_pubchem = '14130')
}
\author{
Zhiwei Zhou
}
