% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools4CommonMetabolomics.R
\name{generateFormulaMassFromSmiles}
\alias{generateFormulaMassFromSmiles}
\title{generateFormulaMassFromSmiles}
\usage{
generateFormulaMassFromSmiles(smiles)
}
\value{
a data.frame with exact_mass and formula
}
\description{
generate formula and mass from smiles
}
\examples{
test <- generateFormulaMassFromSmiles('CN1C=NC(C[C@H](N)C(O)=O)=C1')
}
\author{
Zhiwei Zhou
}
