% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools4CommonMetabolomics.R
\name{limitElements}
\alias{limitElements}
\title{limitElements}
\usage{
limitElements(
  cpd_formula,
  element_necessary = "C",
  element_included = c("C", "H", "N", "O", "P", "S", "F", "Cl", "Br", "I")
)
}
\arguments{
\item{element_necessary}{Necessary elements. Default: 'C'}

\item{element_included}{Limit elements. Default: c("C", 'H', 'N', 'O', 'P', 'S', 'F', 'Cl', 'Br', 'I')}
}
\value{
Logistical value. TRUE or FALSE
}
\description{
limitElements
}
\author{
Zhiwei Zhou
}
