% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools4CommonMetabolomics.R
\name{transformMz}
\alias{transformMz}
\title{transformMz}
\usage{
transformMz(
  exact_mass,
  formula = NULL,
  adduct_list = NULL,
  type = c("adduct", "nl"),
  polarity = c("positive", "negative"),
  ...
)
}
\arguments{
\item{exact_mass}{numeric}

\item{formula}{Default: NULL}

\item{adduct_list}{Default: NULL}

\item{type}{'adduct' or 'nl'.}

\item{polarity}{'positive' or 'negative'}
}
\description{
transformMz
}
\examples{
transformMz(exact_mass = 180.0634, type = 'adduct', polarity = 'positive')
transformMz(exact_mass = 180.0634, type = 'nl', polarity = 'positive')
transformMz(exact_mass = 180.0634, adduct_list = c('[M-H]-', '[M+H]+'))
}
\author{
Zhiwei Zhou
}
